requirements.txt
Flask>=3.0,<4.0
python-dotenv>=1.0
pydantic>=2.0
azure-identity==1.25.0
azure-ai-documentintelligence>=1.0.0
langchain==0.3.27
langchain-openai==0.3.33
langgraph==0.6.7
langgraph-checkpoint==2.1.1
pypdf>=4.0
pdf2image>=1.17
pytesseract>=0.3.10
pandas>=2.0
openpyxl>=3.1
faiss-cpu>=1.8
markdown>=3.5
PyYAML>=6.0
ruff>=0.4
black>=24.0
pytest>=8.0
